+++ src/components/inputs/DateAndTimePicker/DayTimePicker.stories.tsx
@@ -1,4 +1,4 @@
-import { Meta, StoryObj } from "@storybook/react";
+import { Meta, StoryObj } from "@storybook/react-vite";
 import DateTimePicker from "./DateTimePicker";
 import { Box } from "@mui/material";
 import { useState } from "react";