+++ src/theme/colors/GraphOrange.ts
@@ -4,11 +4,13 @@ import { alpha } from "@mui/material/styles";
 const main = "#F2A64B";
 
 const GraphOrange = {
-  primary: {
-    main: main,
-    dark: alpha(main, 0.7),
-    light: alpha(main, 0.5),
-    contrastText: common.black,
+  dark: {
+    primary: {
+      main,
+      dark: alpha(main, 0.7),
+      light: alpha(main, 0.5),
+      contrastText: common.black,
+    },
   },
 };