+++ src/utils/authorize-flow/index.stories.tsx
@@ -1,5 +1,5 @@
 // OAuthFlowDemo.stories.tsx
-import type { Meta, StoryObj } from "@storybook/react";
+import type { Meta, StoryObj } from "@storybook/react-vite";
 import { useState, useEffect } from "react";
 import { QueryClient } from "@tanstack/react-query";
 import AuthServerOAuth2Client, { UserInfo } from "@telicent-oss/fe-auth-lib";